requirements.txt
# SWIFT-LLM Dependencies
# ======================

# Core ML/AI
torch>=2.0.0
transformers>=4.36.0
sentence-transformers>=2.2.0

# Vector Search
faiss-cpu>=1.7.4
chromadb>=0.4.0

# API Clients
openai>=1.0.0
groq>=0.4.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
datasets>=2.14.0

# Evaluation
scikit-learn>=1.3.0
tqdm>=4.66.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
tenacity>=8.2.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0

# Optional: MLX for Apple Silicon (uncomment if on Mac)
# mlx>=0.0.10
# mlx-lm>=0.0.6

